Implement QldbHash
This commit implements QldbHash according to the
specification. QldbHashes are used by the driver to produce a "commit
digest", without which commits cannot be made. They're essentially
just wrappers over SHA256 with a 'dot' operator that can be used to
combine two hashes into one.

The most tricky part is the ordering of hashes, which is well
documented in the implementation.

This commit includes tests that exercise the dot function against
various input permutations.

The Java implementation has constructors that produce QldbHashes from
either an IonValue or String. We will implement these in future
commits, as we pull in our Ion dependencies.