Add an abstraction over QLDBSession
The QLDBSession service defines a single "send command" RPC method,
where the command is a tagged union. For example, a command may
be (exactly one of) StartTransaction, ExecuteStatement, etc.

Rusoto is a SDK for AWS, decomposed into modules. In this commit, we
use the Rusoto modules we need to communicate with QLDBSession. We
abstract over Rusoto such that only proper use of the session API is
possible and, as a side-effect, are decoupled from Rusoto.