Add Ion pretty printing and hash(String)
This commit pollyfills to APIs not yet available in Rust's Ion
ecosystem.

The first is a pretty print method. This takes an Ion reader and
produces a String, which is the pretty Ion text representation of the
value held in the reader. This can be used to pretty print results
coming from QLDB queries. In the future, this method will likely be
pulled direction from ion-rs rather than using ion-c-sys directly.

The second is a Ion has implementation that is specific to Ion strings
using a sha256 hasher. This is useful for producing hash values of
QLDB statement strings. And, because statements can embed values, this
is enough functionality to transact with QLDB. Obviously, we really
want to encourage bind params for security reasons, but this feature
gets us going. When Ion hash is available in Rust, we'll ditch this
method too, which should leave us with an empty file!