Add a session pool
This commit implements a `SessionPool` which holds up to `max`
sessions for re-use. A pool is for a particular ledger, and so will
have a 1:1 relationship with the driver.

Sessions leave the pool by calling `next` (literally: they are gone!)
and are returned to the pool via RAII (`Drop`). This behavior is
coordinated by sending the session token through a channel.

When `next` is called, it first looks at the recv end of the
channel. If it finds a valid session, we're done and hand a RAII guard
out to the client. If not, it makes a new one.

`next` is an async method and may block if:

1. There are no free sessions and `max` has been reached. `next` will
block until a session comes back into the pool.

2. If a new session has to be made, then we need a roundtrip to the
service.

Leaking a session (not dropping the handle) will eventually lead
`next` blocking indefinitely.

I'm unconvinced that this implementation is thread-safe. Maybe it is,
probably it isn't. I haven't really thought about that yet, but I do
know that we want to be able to run on a multi-threaded executor. So I
left a FIXME, such that the pool can be polled by drivers running
concurrently on different threads. I believe this implementation is
safe to check in because we haven't marked anything Sync yet. (And, as
of this commit, there isn't even a driver yet!)