Add the driver
This commit implements the `QldbDriver` and supporting acts. Here's
a high-level overview of what we have:

1. QldbDriver can be built with the QldbDriverBuilder
2. The main API is `transact`, which takes in a function
3. The function has one argument: a Transaction, with it's own API
4. Transactions can execute statements or commit/abort

The driver currently only supports an async API, partially because
that's what Rusuto wants to do under the hood anyways. There is a
BlockingQldbDriver that simply blocks on any futures. However, this
driver still exposes the async-Transaction API. As future work, it'd
be great to have a sync-Transaction API to expose
easy-to-use (blocking) C bindings.

The transact API has some subtleties in the type signature that are
worth flagging. First, the param takes a `Fn(Transaction)` as opposed
to `FnOnce` or `FnMut`. This is because we need to be able to retry
transactions (making 'once' not suitable) and want to encourage
idempotent functions (making 'mut' not suitable). Users *can* work
around idempotency with interior mutability, but that is discouraged.

The next subtly worth calling out is that the Fn must return a special
type that is hard to construct manually. Instead, users use an API on
`Transaction` to commit/abort. These methods take the Transaction by
value, making it hard to incorrectly use the transact method (you need
to call one of them, and you can't call them twice).

The commit digest is a bit of a dance due to ownership. Initially I
tried to have the lambda arg be `&mut Transaction` such that executing
statements could mutate a commit digest in the Transaction. That was a
very sad path to go down, given current limitations in the type
system (notably: GAT support). What I settled on is a model where the
transaction accumulates the digest internally and sends it to the
driver on Drop (via a channel). Beacuse the disposition methods take
the tx by value, this has the neat property of immediately making the
digest available to the driver (which has just await'ed the user fn).

The driver interacts with the pooling and retry functionality to
acquire sessions and retry transactions. Hopefully it does it
correctly because there aren't any meaningful tests yet other than the
example included with this commit, the doctest and my own personal
testing. The reason there aren't yet tests is I'm still thinking
through how I want to go about it. (Spoiler: I don't want mocks.)

A couple of other misc. notes:

1. We only support getting values back as Ion binary
2. Pagination is automatic and blocking (there is no async 'read ahead')
3. The User-Agent includes Rust but it is prepended not appended [1]

1: https://github.com/rusoto/rusoto/pull/1838