Update retry conditions to match the Java driver
fixes #9

This commit updates the `on_err` list of exceptions to retry on. In
particular, we now retry on HTTP 500 and 503 responses as well as any
failure in sending the request (request never leaves the client or
response never comes back).

I believe this makes this driver consistent with the logic in the Java
driver, as noted in the above issue.