Run cargo upgrade and fix
This commit runs `cargo upgrade` which introduces 3 breaking changes:

First, the `gen_range` function from the `rand` crate now takes a range
as a single argument rather than as two params. Good change!

Next, tokio is now at 1.0! We make changes to 1/ the runtime
builder (for the blocking driver), 2/ where we don't need mutable refs
anymore, 3/ move away from try_recv as described in
https://github.com/tokio-rs/tokio/pull/2020 and 4/ replace `delay_for`
with `sleep`.

The main reason for this upgrade, beyond hygiene, is to pick up a new
version of rusoto which we will use shortly.