Implement execution stats
This commit implements execution stats for fully buffered results (which
is the only type we currently support).

As we paginate over results, we accumulate total server side processing
time and IO usage. The fully buffered result then presents those to the
user. The 'main' example shows a usage example. Running it on my machine
results in:

```
> RUST_LOG=info cargo run --example main
[2021-01-14T21:46:15Z INFO  main] Creating a QLDB driver
[2021-01-14T21:46:15Z INFO  main] Transaction example 1 now running
[2021-01-14T21:46:16Z INFO  main] Tx 1 returned 1 result(s):
[2021-01-14T21:46:16Z INFO  main] 42
[2021-01-14T21:46:16Z INFO  main] Statement executed in 14ms and used 1 read IOs
[2021-01-14T21:46:16Z INFO  main] Goodbye!
```

There are two things I'm not super excited about in this commit. First,
StatementResults leaks the Rusoto types directly. I think that's a
mistake. Second, it yields `Option` to customers. While this is accurate
given the wire model, the optional values present unneccessary
complexity to users. The Java commit [1] handles this by returning 0s to
customers, which seems right to me.

Regardless, I think this commit adds value and should be merged as is.
Given we're not committing to a stable API yet, this is a pure win.

[1] https://github.com/awslabs/amazon-qldb-driver-java/commit/4cb70843d45330b4884cb21b01e84433ea627359