Handle previously failed StartTransaction attempts
Before this commit, a failure on StartTransaction would return to the
user as a BadRequest. Then, the session would go back into the pool and
immediately be re-used. No good!

Now, `BadRequest` on start session will immediately retry. This isn't
great either (it lacks in safety), but I'm also not sure that plumbing
it down the normal path of retries (sleep, max attempts) is the right
one.

Finally, I tried pretty hard to refactor this method but gave up after 3
attempts. Clearly, something needs to be done to untangle it, but it's
rather tricky.

Ultimately, I figured it's best to get this fix (clearly better behavior
in) and I'll take another shot at the cleanup in a future commit.