Extract start_transaction to Transaction::start
Before this commit, Transaction had a value constructor which wasn't
super useful. The responsibility for construction was split between said
constructor and the caller.

This commit cleans up that split responsibility (e.g. the setup of the
ch pair). Not a huge win, but a step in the right direction.