Start capturing overall stats
Now that starting a transaction is behind `Transaction::start` it makes
it a bit easier to setup the overall accumulated stats for the
transaction. This commit does a few things:

1. The Transaction will have stats that accumulate for all APIs called.
We set those up in `start` and include the StartTransaction stats.

2. We start moving the `accumulate` infrastructure over from Rusoto
types to our types. Much cleaner! The other stuff will be removed
shortly.

In the next commit I plan on doing some cleanup, then after that I will
continue to accumulate stats.