Make the driver cloneable to enable concurrent use
This is a fairly large commit. Before, the driver was none of Send, Sync
or Clone. There was no way to use it in a concurrent application. Your
only choice was a single-threaded Tokio runtime.

There are really three things going on with the driver. First, we have
the client (a Rusoto client) that manages HTTPS connections and
credentials. Next, we have the pool of sessions. And, finally, the
transactional code that customers write.

Before this commit, you *could* make multiple drivers and use the same
Rusoto client (since cloning it is just an Arc clone), but you'd not use
the same SessionPool. In this commit, we make SessionPool cloneable by
reimplementing it on top of bb8.

That leaves us with the challenge of multiple threads using the same
driver. In essence, the top-level requirement that tokio has on the
Future being Send created a lot of trouble. I added Arcs and wove
through trait requirements until everything eventually worked.

There was a particular question around TransactionRetryPolicy - if there
is only 1 driver with shallow clones, then what are the Send/Sync
requirements on the policy? Ultimately, I decided to keep the
implementation requirements of the policy simple and wrap it in a Mutex.
That bubbled up yet more Send requirements, leading me to change the
trait itself. Now it's not async_trait and the sleep is done by the
caller not the code itself. I think this is better anyways, the previous
"policy can sleep" strategy was going to make the Mutex tricky to use.

I added some testing, which resulted in weaving a `C: QldbSession`
throughout the codebase, as well as some changes to the builder. There
is a hand-written testing client that isn't very good. A new test shows
how you can use the driver from more than one thread. The test doesn't
pass 100% of the time; which is a good thing at some level! Bear with
me. The mock returns results in order, while the tests run in OS
scheduling order. So, tx-1 might get the mock results for tx-0. The
invariant checking the driver does causes the failure. As future work,
we should make the mock smarter so that we can control more specifically
what the return order is.

Also: move from std::result to anyhow::Result (removes box StdError
type) and start trying to rationalize how the driver will support real
threads concurrency.