Rename ok to commit
Initially it seemed like a good idea to call the method 'ok' in the
style of Ok/Err Result types. However, when explaining usage to users, I
landed up having to explain that "ok calls commit".

This commit deprecates `ok` so that we can stage the change in one
commit here, then upgrade the shell in a second commit.