Refactor commit/abort from driver -> transaction
In the early design of the driver, there was no explicit commit/abort
API that the user had to call, e.g. this was valid usage:

```
let value = driver.transact({|_tx| 42});
```

This presented several design challenges:

1. The lambda argument had to move the value in, due to various
limitations around lifetimes in async rust (see a previous commit
message for a long explanation).

2. But.. of course we need a way to know what the commit digest is (to
actually commit the transaction). This would be easy to do if the lambda
returned the transaction value back to us, but then users can't export
results to their code.

The hack I came up with was to use a channel that sent the digest as
part of its drop implementation. However, later we landed up having an
explicit commit API anyways, so now it seems like all this channel
nonsense isn't worth it.

This really came to a head when trying to finish up execution stats and
having to also send the as-accumulated stats back over the channel. The
real rub was it was never going to be possible for the user to capture
the stats of *just* the commit call (since those are hidden away in the
implementation). Also, it's kind of weird that if you measure the call
to commit it would appear to take 0ms.

With this change, calling `tx.commit` *actually sends the commit*.