Further refactor the `abort` API
As per the comments, the abort API allowed returning data (bad) and
showed up as an `Ok` variant to the caller (bad). Now, calling abort
does *not* let the caller export data and, instead, the abort will cause
the `transact` call to return an `Err`.

This (aborted error) is an `anyhow::Result`, so if application code
wants to specifically handle it, they must use the downcast method as
shown in a test.

I think this change makes it easier to write correct code and also
removes the need to put fake `R` data in the call to abort.