Refactor the API into the pooled connection
As per the comment I added earlier and have now removed..

This commit ensures that any API call that gets an InvalidSession will
prevent the session ("connection") from going back into the pool. This
removes a bunch of places in the Driver where we had to do these
checks (win).