Change the build method to be async
Before this commit, the build function was *not* async and would use the
`build_unchecked` method of bb8. This method actually fails if you're
not in a tokio runtime, so that's bad. Further, it turns out that it's
really hard to know why things are failing in bb8 because the `get`
method only ever returns a connect or a bb8 timeout error.

With this commit, we use the `async build()` method which will fail if
the minimum number of connections aren't established. This makes it more
obvious you actually needed a tokio runtime and propagates the last
connection error if one couldn't be established within the timeout.

The default timeout is lowered from 30s to 10s. I didn't want to make it
low enough that a spotty connection would make all use-cases invalid,
but even 10s seems like forever these days..