Refactor QldbHttp1Connection to lower level
This commit changes the wrapper to use the lower level of
abstraction (the rusoto QldbSession trait rather than the QldbSessionApi
trait we define). The benefit of this change is we don't need to keep up
with every command variant; the check is now generic over `SendCommand`.

The only downside is I had to change a `QldbError::IllegalState` to a
panic, which I think is fine for now.