Demonstrate & enable functions as transact args
This commit includes an example of how to use a `fn` as an argument to
the driver's `transact` method.

I suspect this will be a common way to write code in larger systems as
it allows for the business logic inside the function to be easily tested
as it is naturally separted from other flow (initial reads, calling
`driver.transact`, etc.) It also allows for normal function
decomposition.

I'm not happy with how we leak the `C` type. I suspect we're going to
have to do some internal boxxing.