Reimplement execution stats off the smithy model
Before this commit, we had an in-crate defined struct for
ExecutionStats. This let us decouple from the rusoto type and have a
place to accumulate stats over the course of a transaction. Now, we use
the smithy-generated types!

The way this works is that the -core package now depends on
aws-sdk-qldbsession with the "client" feature disabled. This means we
get the shapes, but not the networking dependencies.

For gravy, a macro now takes care of implementing some of the conversion
code. Eventually, the stuff in the `rusoto_support` module will move out
of the -core crate.