Remove last usages of rusoto in the driver
Before this commit, there was no actual way to use the awssdk with the
driver (other than how the shell was using it). This commit moves the
abstraction the shell was using into this crate and moves away from
using the rusoto credential providers in the example. Now, we can run
the examples using only the awssdk and not use rusoto at all!

Ideally, we should have -core not generate the client such that -core
remains sdk agnostic. I'll fix that in a later commit.

The awssdk was update to alpha19.