Reimplement the driver using event streaming
There are several broad strokes made in this commit:

1. QldbSession v1 is gone entirely. I initially made an attempt to have
these live side-by-side, but they're just so different. If there is a
desire to allow runtime switching of underlying drivers, we can bring
that back in a future commit.

2. In particular, having QldbError::SdkError was a huge mistake. Going
forward, we're going to abstract the SDK errors entirely rather than
couple the transport and application layers.

3. Eventstreaming is used in a 1:1 fashion only - 1 request, 1 response.
In the future, we may have concurrent request-responses, but we will
need correlation ids to support that.

4. I've entirely punted on the user errors and retries for now.

5. The support for dynamic SDKs has been removed. I think it's useful to
get this back, but I want to try get it done using the Smithy patterns
rather than the trait stuff. As a result, a bunch of indirection is gone
which made this much easiser to write.

Of note, I really like how the `<C>` parameter just went away -
especially in the user-facing `TransactionAttempt<C>`. This is really
nice.

I was also pleased how the bb8 pool abstraction actually made a very
natural place to stash the input/output channels.

There is no testing yet.