Remove core crate
This commit removes the distinction between the -core crate and the
driver facade.

Initially the idea was to have a "pure" implementation that didn't have
any networking stuff so that we could compile to wasm and dependency
inject different SDKs. However, we're not actually doing that yet and
this indirection wasn't helping us move quickly.

Furthermore, the new SDK has taken the directly where the toplevel
client isn't a trait. Rather, SmithyConnector (1 layer down) allows for
pluggable connectivity. This is really nice, and means that when we
re-pursue the wasm directly we might not even need this two-crate
approach anyway!